brandQty,categoryQty

+++ src/main/java/com/backend/backend_java/entity/Brand.java
@@ -23,7 +23,7 @@ public class Brand {
     @Size(min = 2, message = "Brand name must contain at least 2 characters")
     private String brandName;
 
-    private Integer brandQty; // Số lượng sản phẩm thuộc brand
+    private Long brandQty; // Số lượng sản phẩm thuộc brand
 
     @OneToMany(mappedBy = "brand", cascade = CascadeType.ALL)
     private List<Product> products;
